Chore: Ajax 요청 이후 setState 문제 해결을 위한 진단 코드 작성
